powerpc/powernv: Initialise nest mmu
POWER9 contains an off core mmu called the nest mmu (NMMU). This is used
by other hardware units on the chip to translate virtual addresses into
real addresses. The unit attempting an address translation provides the
majority of the context required for the translation request except for
the base address of the partition table (ie. the PTCR) which needs to be
programmed into the NMMU.

This patch adds a call to OPAL to set the PTCR for the nest mmu in
opal_init().

Signed-off-by: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>