mm/migrate: Add a unmap and pin migration mode
Some drivers do page migration only to ensure that the device has
exclusive access to a particular page. This is required to allow a
driver to implement atomic access for example. In these cases it is
advantageous to remap the page for device access without actually
migrating the data.

To allow this kind of access introduce an unmap and pin flag called
MIGRATE_PFN_PIN/UNPIN for migrate pfns. This will cause the original
page to be remapped to the provided device private page as normal, but
instead of returning the original CPU page to the system it will pin it
and leave it isolated from the LRU.

This ensures the page remains pinned so that a device may access it
exclusively. Any userspace CPU accesses will fault and trigger the
normal device private migrate_to_ram() callback which must migrate the
mapping back to the original page, after which the device will no longer
have exclusive access to the page.

Signed-off-by: Alistair Popple <apopple@nvidia.com>