nouveau/svm: Implement atomic SVM access
NVIDIA GPUs do not support direct atomic access to system memory via
PCIe. Instead this must be emulated by granting the GPU exclusive access
to the memory. This is achieved by migrating the userspace mappings to
device private pages whilst leaving the actual page in place.

The driver then grants the GPU permission to update the page undergoing
atomic access via the GPU page tables. When CPU access to the page is
required a CPU fault is raised which calls into the device driver via
migrate_to_ram() to remap the page into the CPU page tables and revoke
GPU access.

As this does not require the page data to migrate it is safe to continue
when other kernel users hold a reference on the page. Only anonymous
private memory is supported for SVM and these users will not write to
the page nor expect updates to be atomic.

Signed-off-by: Alistair Popple <apopple@nvidia.com>