mm: Introduce struct vm_account for pins cgroup
Accounting pinned memory to a cgroup requires callers to keep a
reference to the pins_cgroup structure until the pages are
unpinned. Pinning also requires a reference to mm_struct for
statistics in mm->pinned_vm.

To make this easier introduce a vm_account structure which can be
embedded in driver contexts to store these references. This struct
then be passed into new versions of {un}account_pinned_vm() to account
pins to both mm->pinned_vm and the pins_cgroup.

As not all drivers can be easily converted the old functions are
renamed to __{un}account_pinned_vm(). Once all drivers are converted
to use the vm_account structure these functions will be removed.

Signed-off-by: Alistair Popple <apopple@nvidia.com>