mm: Remove uses of PFN_DEV
PFN_DEV is used by callers of dax_direct_access() to figure out if the
returned PFN is associated with a page or not. However all DAX PFNs
now require an assoicated ZONE_DEVICE page so can assume a page
exists.

Other users of PFN_DEV were setting it before calling
vmf_insert_mixed(). This is unnecessary as it is no longer checked,
instead relying on pfn_valid() to determine if there is an associated
page or not.

Signed-off-by: Alistair Popple <apopple@nvidia.com>